requirements.txt
python-gitlab==1.7.0